[REFACTOR] import문 오류 수정 (#75)
